feat(order): complete order manage without image

+++ apps/api/src/services/cart.service.ts
@@ -1,8 +1,11 @@
 import { Request } from 'express';
 import { prisma } from '@/libs/prisma';
 import { z } from 'zod';
-import { deleteCartSchema, upsertCartSchema } from '@/libs/zod/cartSchema';
-import { BadRequestError } from '@/utils/error';
+import {
+  deleteCartSchema,
+  upsertCartSchema,
+} from '@/libs/zod-schemas/cart.schema';
+import { AuthError, BadRequestError } from '@/utils/error';
 
 export class CartService {
   async getCartByUserId(req: Request) {
@@ -13,6 +16,8 @@ export class CartService {
 
   async upsertCart(req: Request) {
     const user_id = req.user.id;
+    if (!user_id || req.user.role != 'customer')
+      throw new AuthError('not authorized');
     const { quantity, store_stock_id } = req.body as z.infer<
       typeof upsertCartSchema
     >;
@@ -38,8 +43,8 @@ export class CartService {
         quantity: Number(quantity),
       },
       create: {
-        user_id: '',
-        store_stock_id: '',
+        user_id,
+        store_stock_id,
         quantity: quantity,
       },
     });
@@ -47,9 +52,16 @@ export class CartService {
 
   async deleteProductInCart(req: Request) {
     const user_id = req.user.id;
+    if (!user_id || req.user.role != 'customer')
+      throw new AuthError('not authorized');
     const { store_stock_id } = req.body as z.infer<typeof deleteCartSchema>;
     return prisma.cart.delete({
-      where: { user_id_store_stock_id: { user_id, store_stock_id } },
+      where: {
+        user_id_store_stock_id: {
+          user_id,
+          store_stock_id,
+        },
+      },
     });
   }
 }
